Added random for method for flat and bounded distributions.
