+++ pymc3/distributions/distribution.py
@@ -2,6 +2,9 @@ import theano.tensor as t
 import numpy as np
 from ..model import Model
 
+from theano import function
+from ..model import get_named_nodes
+
 __all__ = ['DensityDist', 'Distribution', 'Continuous', 'Discrete', 'NoDistribution', 'TensorType']
 
 
@@ -96,3 +99,57 @@ class MultivariateContinuous(Continuous):
 class MultivariateDiscrete(Discrete):
 
     pass
+
+def draw_values(params, point=None):
+    """
+    Draw (fix) parameter values. Handles a number of cases:
+
+        1) The parameter is a scalar
+        2) The parameter is an *RV
+
+            a) parameter can be fixed to the value in the point
+            b) parameter can be fixed by sampling from the *RV
+            c) parameter can be fixed using tag.test_value (last resort)
+
+        3) The parameter is a tensor variable/constant. Can be evaluated using
+        theano.function, but a variable may contain nodes which
+
+            a) are named parameters in the point
+            b) are *RVs with a random method
+
+    """
+    # Distribution parameters may be nodes which have named node-inputs
+    # specified in the point. Need to find the node-inputs to replace them.
+    givens = {}
+    for param in params:
+        if hasattr(param, 'name'):
+            named_nodes = get_named_nodes(param)
+            if param.name in named_nodes:
+                named_nodes.pop(param.name)
+            for name, node in named_nodes.items():
+                givens[name] = (node, draw_value(node, point=point))
+    values = [None for _ in params]
+    for i, param in enumerate(params):
+        # "Homogonise" output
+        values[i] = np.atleast_1d(draw_value(param, point=point, givens=givens.values()))
+    if len(values) == 1:
+        return values[0]
+    else:
+        return values
+
+def draw_value(param, point=None, givens={}):
+    if hasattr(param, 'name'):
+        if hasattr(param, 'model'):
+            if point is not None and param.name in point:
+                return point[param.name]
+            elif hasattr(param, 'random') and param.random is not None:
+                return param.random(point=point, size=None)
+            else:
+                return param.tag.test_value
+        else:
+            return function([], param,
+                            givens=givens,
+                            rebuild_strict=False,
+                            allow_input_downcast=True)()
+    else:
+        return param
